+++ atlasdb-dagger/src/main/java/com/palantir/atlasdb/services/test/TestSweeperModule.java
@@ -22,10 +22,15 @@ import javax.inject.Named;
 import javax.inject.Singleton;
 
 import com.google.common.collect.ImmutableList;
+import com.palantir.atlasdb.AtlasDbConstants;
 import com.palantir.atlasdb.cleaner.Follower;
 import com.palantir.atlasdb.keyvalue.api.KeyValueService;
+import com.palantir.atlasdb.persistentlock.KvsBackedPersistentLockService;
+import com.palantir.atlasdb.persistentlock.NoOpPersistentLockService;
+import com.palantir.atlasdb.persistentlock.PersistentLockService;
 import com.palantir.atlasdb.services.ServicesConfig;
 import com.palantir.atlasdb.sweep.CellsSweeper;
+import com.palantir.atlasdb.sweep.PersistentLockManager;
 import com.palantir.atlasdb.sweep.SweepTaskRunner;
 import com.palantir.atlasdb.transaction.impl.SerializableTransactionManager;
 import com.palantir.atlasdb.transaction.impl.SweepStrategyManager;
@@ -58,6 +63,18 @@ public class TestSweeperModule {
         this.immutableTs = immutableTs;
     }
 
+
+    @Provides
+    public PersistentLockManager providePersistentLockManager(@Named("kvs") KeyValueService kvs,
+            ServicesConfig config) {
+        PersistentLockService persistentLockService = kvs.supportsCheckAndSet() ?
+                KvsBackedPersistentLockService.create(kvs, config.atlasDbConfig().initializeAsync()) :
+                new NoOpPersistentLockService();
+        return new PersistentLockManager(
+                persistentLockService,
+                AtlasDbConstants.DEFAULT_SWEEP_PERSISTENT_LOCK_WAIT_MILLIS);
+    }
+
     @Provides
     @Singleton
     public SweepTaskRunner provideSweepTaskRunner(SerializableTransactionManager txm,
@@ -65,6 +82,7 @@ public class TestSweeperModule {
                                                   TransactionService transactionService,
                                                   SweepStrategyManager sweepStrategyManager,
                                                   Follower follower,
+                                                  PersistentLockManager persistentLockManager,
                                                   ServicesConfig config) {
         LongSupplier unreadable = unreadableTs.orElse(txm::getUnreadableTimestamp);
         LongSupplier immutable = immutableTs.orElse(txm::getImmutableTimestamp);
@@ -74,7 +92,11 @@ public class TestSweeperModule {
                 immutable,
                 transactionService,
                 sweepStrategyManager,
-                new CellsSweeper(txm, kvs, ImmutableList.of(follower), config.atlasDbConfig().initializeAsync()));
+                new CellsSweeper(
+                        txm,
+                        kvs,
+                        persistentLockManager,
+                        ImmutableList.of(follower)));
     }
 
 }